added validations for comments contrller

+++ server/controllers/comment-ctrl.js
@@ -4,7 +4,14 @@ const commentCollection = require('../services/comment-service');
 module.exports = (socket) => {
 
     socket.on('create-comment', (commentDetails) => {
-
+        if (commentDetails && (!commentDetails.description || !commentDetails.postedBy || !commentDetails.post)) {
+            return socket.emit('new-comment-added', {
+                success: false,
+                data: {
+                    "message": 'Required fields not present. Check your payload'
+                }
+            });
+        }
         commentCollection.insert(commentDetails, (err, response) => {
             if (err) {
                 socket.emit('new-comment-added', {
@@ -21,6 +28,15 @@ module.exports = (socket) => {
     });
 
     socket.on('update-comment', (commentDetails) => {
+
+        if (commentDetails && (!commentDetails.description || !commentDetails._id)) {
+            return socket.emit('comment-updated', {
+                success: false,
+                data: {
+                    "message": 'Required fields not present. Check your payload'
+                }
+            });
+        }
 
         commentCollection.update(commentDetails, (err, response) => {
             if (err) {
@@ -39,6 +55,15 @@ module.exports = (socket) => {
 
     socket.on('delete-comment', (commentDetails) => {
 
+        if (commentDetails && (!commentDetails.postID || !commentDetails.id)) {
+            return socket.emit('comment-delete', {
+                success: false,
+                data: {
+                    "message": 'Required fields not present. Check your payload'
+                }
+            });
+        }
+
         commentCollection.delete(commentDetails, (err, response) => {
             if (err) {
                 socket.emit('comment-delete', {
@@ -55,6 +80,14 @@ module.exports = (socket) => {
     });
 
     socket.on('fetch-all-comments', (PostID) => {
+        if (!PostID) {
+            return socket.emit('comments-fetched', {
+                success: false,
+                data: {
+                    "message": 'Send Post ID to fetch all comments'
+                }
+            });
+        }
         commentCollection.fetchPostComments(PostID, (err, response) => {
             if (err) {
                 socket.emit('comments-fetched', {
